requirements.txt
# KnowLigo - Dependencies
# Requiere: Python 3.11 o 3.12 (wheels pre-compilados disponibles)

# Web framework
fastapi==0.115.0
uvicorn[standard]==0.32.1
pydantic==2.9.2
pydantic-settings==2.12.0
requests==2.32.3
httpx==0.27.2
python-dotenv==1.0.1
markdown==3.7

# LLM API
groq==0.11.0

# RAG dependencies
faiss-cpu==1.13.2
sentence-transformers==3.3.1
rank-bm25>=0.2.2

# Testing
pytest==8.3.5
pytest-asyncio==0.25.3

# Note: sqlite3 viene built-in con Python, no requiere instalación
# Note: sentence-transformers trae como dependencias:
#   - torch, transformers, numpy, scipy, scikit-learn, etc.
#   - Usa wheels pre-compilados cuando están disponibles
